# Containers-kuber8,Docker,Openshit/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Product Info",
    "description": "A sample API that uses Product Information microservice to provide CRUD operations"
  },
  "host": "--Replace this with your host URL--",
  "basePath": "/",
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/products": {
      "get": {
        "description": "Returns all products",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all products",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          }
        }
      },
      "post": {
        "description": "Creates a new product",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "description": "Product to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Product creation response"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "description": "Returns a product based on a product id",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of product to fetch",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Product response",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      },
      "put": {
        "description": "Updates a product based on a product id",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of product to fetch",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "product",
            "in": "body",
            "description": "Product details to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateProduct"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product updated"
          }
        }
      },
      "delete": {
        "description": "deletes a single product based on the product id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of product to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted"
          }
        }
      }
    }
  },
  "definitions": {
    "Product": {
      "type": "object",
      "allOf": [
        {
          "required": ["id", "name", "price"],
          "properties": {
            "id": {
              "type": "integer"
            },
            "name": {
              "type": "string"
            },
            "price": {
              "type": "number"
            }
          }
        }
      ]
    },
    "UpdateProduct": {
      "type": "object",
      "allOf": [
        {
          "properties": {
            "name": {
              "type": "string"
            },
            "price": {
              "type": "number"
            }
          }
        }
      ]
    }
  }
}